penge-block-layout: Add a new block layout manager
This provides a layout manager we can use along with a ClutterBox to give us a
"block layout". The size of the blocks is determined based on a minimum tile
size and how many could fit across. Items in the layout can have a different
number of column spans covering the wholes in the layout.