events-pane: Don't paint skipped children
Since in the events pane we only want to show children we have space for we
have to be cunning we have to have a custom allocate. When we have allocated
as many as we can we allocate those left to 0,0 with height and width 0,0 and
then skip those in a custom paint function.

This fixes: http://bugzilla.meego.com/show_bug.cgi?id=490 and
http://bugzilla.meego.com/show_bug.cgi?id=472